refactor: code 가독성 있게 수정
