--- src/pages/FriendPage/add-friend/index.tsx
@@ -1,5 +1,0 @@
-import * as styles from './styles.css';
-
-export const AddFriend = () => {
-  return <button className={styles.addFriend} />;
-};
