--- src/pages/FriendPage/add-friend/styles.css.ts
@@ -1,9 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-import { rem } from '@/styles';
-
-export const addFriend = style({
-  width: rem(56),
-  height: rem(46),
-  background: "url('/assets/images/add-friend.svg') center/contain no-repeat",
-});
