--- src/pages/FriendPage/alarm/index.tsx
@@ -1,5 +1,0 @@
-import * as styles from './styles.css';
-
-export const Alarm = () => {
-  return <button className={styles.alarm} />;
-};
