+++ src/pages/SignUpPage/styles.css.ts
@@ -60,7 +60,7 @@ export const verifyButton = style({
   height: rem(40),
   color: theme.color.white,
   fontSize: rem(18),
-  background: `url('/assets/images/base-button-on.png') no-repeat center`,
+  background: `url('/assets/images/base-button-normal.png') no-repeat center`,
   backgroundSize: 'contain',
   opacity: 1,
   transition: 'opacity 250ms ease',
@@ -80,7 +80,7 @@ export const submitButton = style({
   marginTop: rem(20),
   color: theme.color.white,
   fontSize: rem(22),
-  background: `url('/assets/images/base-button-on.png') no-repeat center`,
+  background: `url('/assets/images/base-button-normal.png') no-repeat center`,
   backgroundSize: 'contain',
   opacity: 1,
   transition: 'opacity 250ms ease',
